Send configuration object to Benchmark child processes instead of dictionary. Refactoring of diart.models
