+++ src/diart/blocks/__init__.py
@@ -13,10 +13,6 @@ from .embedding import (
     OverlapAwareSpeakerEmbedding,
 )
 from .segmentation import SpeakerSegmentation
-from .diarization import (
-    OnlineSpeakerDiarization,
-    BasePipeline,
-    PipelineConfig,
-    BasePipelineConfig,
-)
+from .diarization import OnlineSpeakerDiarization, BasePipeline
+from .config import BasePipelineConfig, PipelineConfig
 from .utils import Binarize, Resample, AdjustVolume
